feat: amelioration de l'achitecture du code + ajout de loading au premier chargement ainsi que lors d'une erreur de trajet
